fix tests and conversion
